Refactor security services out of main
